explicitly require beaker-puppet
This is a transitive dependency. In the past we pulled it in via beaker
itself. Since the recent beaker/beaker-rspec we need to require it
explicitly. The new beaker version also requires us to set
BEAKER_IS_PE=no for each acceptance test, otherwise it tries to install
Puppet Enterprise...